pyproject.toml
[tool.poetry]
name = "analysis-age-gender-cvae"
version = "0.1.0"
description = ""
authors = ["mitsu-h <mitsui.engineer@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytorch-lightning = "^2.1.2"
ipython = "^8.17.2"
pillow = "^10.1.0"
pandas = "^2.1.3"
torchvision = "^0.16.1"
mlflow = "^2.8.1"
ipykernel = "^6.26.0"

[tool.poetry.group.dev.dependencies]
plotly = "^5.18.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
